Use identity from core context
