+++ src/lib/components/workflow/client-actions/batch-cancel-confirmation-modal.svelte
@@ -12,7 +12,7 @@
     BATCH_OPERATION_CONTEXT,
     type BatchOperationContext,
   } from '$lib/pages/workflows-with-new-search.svelte';
-  import { batchCancelWorkflows } from '$lib/services/batch-service';
+  import { batchCancelWorkflows } from '$lib/services/workflow-batch-service';
   import { toaster } from '$lib/stores/toaster';
   import { workflowsQuery } from '$lib/stores/workflows';
   import { getIdentity } from '$lib/utilities/core-context';
@@ -60,6 +60,7 @@
     error = '';
     const options = {
       namespace,
+      identity,
       reason: $reason ? $reason : reasonPlaceholder,
       jobId: $jobId || jobIdPlaceholder,
       ...($allSelected
