+++ src/lib/components/workflow/client-actions/batch-terminate-confirmation-modal.svelte
@@ -12,7 +12,7 @@
     BATCH_OPERATION_CONTEXT,
     type BatchOperationContext,
   } from '$lib/pages/workflows-with-new-search.svelte';
-  import { batchTerminateWorkflows } from '$lib/services/batch-service';
+  import { batchTerminateWorkflows } from '$lib/services/workflow-batch-service';
   import { toaster } from '$lib/stores/toaster';
   import { workflowsQuery } from '$lib/stores/workflows';
   import { getIdentity } from '$lib/utilities/core-context';
@@ -61,6 +61,7 @@
     try {
       const options = {
         namespace,
+        identity,
         reason: $reason ? $reason : reasonPlaceholder,
         jobId: $jobId || jobIdPlaceholder,
         ...($allSelected
